+++ demo/src/main/res/layout/activity_color_picker.xml
@@ -31,7 +31,7 @@
       android:layout_weight="1"
       android:padding="25dp">
 
-    <com.jrummyapps.android.colorpicker.view.ColorPickerView
+    <com.jrummyapps.android.colorpicker.ColorPickerView
         android:id="@id/colorpickerview__color_picker_view"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
@@ -46,7 +46,7 @@
         android:layout_marginTop="16dp"
         android:orientation="horizontal">
 
-      <com.jrummyapps.android.colorpicker.view.ColorPanelView
+      <com.jrummyapps.android.colorpicker.ColorPanelView
           android:id="@id/colorpickerview__color_panel_old"
           android:layout_width="0px"
           android:layout_height="fill_parent"
@@ -61,7 +61,7 @@
           android:text="→"
           android:textSize="20sp"/>
 
-      <com.jrummyapps.android.colorpicker.view.ColorPanelView
+      <com.jrummyapps.android.colorpicker.ColorPanelView
           android:id="@id/colorpickerview__color_panel_new"
           android:layout_width="0px"
           android:layout_height="fill_parent"
