+++ demo/src/main/res/xml/main.xml
@@ -18,7 +18,7 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
 
   <PreferenceCategory android:title="Demo">
-    <com.jrummyapps.android.colorpicker.preference.ColorPreference
+    <com.jrummyapps.android.colorpicker.ColorPreference
         android:defaultValue="0xFF000000"
         android:key="color"
         android:summary="Show our color picker using the custom ColorPreference class."
