+++ library/src/main/java/com/jrummyapps/android/colorpicker/DrawingUtils.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2015. Jared Rummler <jared.rummler@gmail.com>
- * Copyright (C) 2015 Daniel Nilsson
+ * Copyright (C) 2017 JRummy Apps Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +12,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
+ * Raw
  */
 
-package com.jrummyapps.android.colorpicker.view;
+package com.jrummyapps.android.colorpicker;
 
 import android.content.Context;
 import android.util.DisplayMetrics;
